Update server.py
New , clean version of server. Had to comment out the HTTPS stuff to get the raw socket part to work for testing. I'll be working from this version to add the Telnet stuff.
Coleman - All I did was comment out the HTTPS stuff, I didn't modify it. If you need to work on those parts for testing you should just be able to uncomment those sections.