Update report.py
Well that's it.
Removed some of the testing only stuff to make it functional.
This gets called within the HTTP(S) handler in the server module.
Couldn't get payload generator to run on my machine so I wasn't able to look at the file structure and naming conventions to get the counting part to work correctly. 
Since we are only sending one file in the demo, I have hard coded the report.txt writing part to print 1 success out of 1. We can fix it later today if we want to. Before sending the real code to the clients.